requirements.txt
pandas
numpy
matplotlib
seaborn
pytesseract
streamlit
opencv-python
scikit-learn
tensorflow
easyocr
python-dotenv
mysql-connector-python
requests
secure-smtplib
pillow
azure-ai-formrecognizer
SQLAlchemy
PyMySQL
langchain
langchain-community
langchain-groq
langchain-openai
suno-bark
scipy
ffmpeg-python
openai
moviepy
langchain_openai
langchain_core
OpenAI
litellm
chromadb
PyPDF2
PyMuPDF
flask
fitz
langchain_experimental
tensorflow
streamlit-authenticator
firebase-admin
langchain-google-genai 
google-generativeai
pypdf
landingai